Move the logic into appdetect and surface metadata that way
