+++ cli/azd/internal/appdetect/appdetect.go
@@ -161,10 +161,6 @@ func (p *Project) HasWebUIFramework() bool {
 	return false
 }
 
-type Docker struct {
-	Path string
-}
-
 type projectDetector interface {
 	Language() Language
 	DetectProject(ctx context.Context, path string, entries []fs.DirEntry) (*Project, error)
