+++ cli/azd/internal/appdetect/docker.go
@@ -1,7 +1,10 @@
 package appdetect
 
 import (
+	"bufio"
+	"fmt"
 	"io/fs"
+	"os"
 	"path/filepath"
 	"strings"
 )
@@ -16,4 +19,32 @@ func detectDocker(path string, entries []fs.DirEntry) (*Docker, error) {
 	}
 
 	return nil, nil
+}
+
+func detectPortInDockerfile(
+	filePath string) []int {
+	file, err := os.Open(filePath)
+	if err != nil {
+		return []int{}
+	}
+	defer file.Close()
+
+	var result []int
+	scanner := bufio.NewScanner(file)
+	for scanner.Scan() {
+		line := scanner.Text()
+		if strings.HasPrefix(line, "EXPOSE") {
+			var port int
+			_, err := fmt.Sscanf(line, "EXPOSE %d", &port)
+			if err == nil {
+				result = append(result, port)
+			}
+		}
+	}
+	return result
+}
+
+type Docker struct {
+	Path         string
+	ExposedPorts []int
 }
