+++ cli/azd/internal/appdetect/testdata/Dockerfile/DockerfileNoPort
