+++ cli/azd/internal/appdetect/testdata/Dockerfile/DockerfilePort3100And80
