+++ cli/azd/internal/repository/infra_confirm.go
@@ -1,10 +1,8 @@
 package repository
 
 import (
-	"bufio"
 	"context"
 	"fmt"
-	"os"
 	"path/filepath"
 	"regexp"
 	"strconv"
@@ -104,12 +102,11 @@ func (i *Initializer) infraSpecFromDetect(
 		if svc.Docker == nil || svc.Docker.Path == "" {
 			// default builder always specifies port 80
 			serviceSpec.Port = 80
-
 			if svc.Language == appdetect.Java {
 				serviceSpec.Port = 8080
 			}
 		} else {
-			ports := i.detectPortInDockerfile(svc.Docker.Path)
+			ports := svc.Docker.ExposedPorts
 			if len(ports) == 1 {
 				serviceSpec.Port = ports[0]
 			}
@@ -199,27 +196,4 @@ func (i *Initializer) infraSpecFromDetect(
 	}
 
 	return spec, nil
-}
-
-func (i *Initializer) detectPortInDockerfile(
-	filePath string) []int {
-	file, err := os.Open(filePath)
-	if err != nil {
-		return []int{}
-	}
-	defer file.Close()
-
-	var result []int
-	scanner := bufio.NewScanner(file)
-	for scanner.Scan() {
-		line := scanner.Text()
-		if strings.HasPrefix(line, "EXPOSE") {
-			var port int
-			_, err := fmt.Sscanf(line, "EXPOSE %d", &port)
-			if err == nil {
-				result = append(result, port)
-			}
-		}
-	}
-	return result
 }
