+++ cli/azd/internal/repository/infra_confirm_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"os"
-	"path/filepath"
 	"strings"
 	"testing"
 
@@ -225,33 +224,4 @@ func TestInitializer_infraSpecFromDetect(t *testing.T) {
 			require.Equal(t, tt.want, spec)
 		})
 	}
-}
-
-func TestDetectPortInDockerfile(t *testing.T) {
-	i := &Initializer{
-		console: input.NewConsole(
-			false,
-			false,
-			input.Writers{Output: os.Stdout},
-			input.ConsoleHandles{
-				Stderr: os.Stderr,
-				Stdin:  os.Stdin,
-				Stdout: os.Stdout,
-			},
-			nil,
-			nil),
-	}
-	var port []int
-	port = i.detectPortInDockerfile(filepath.Join("testdata", "Dockerfile", "DockerfilePort80"))
-	require.Equal(t, 1, len(port))
-	require.Equal(t, 80, port[0])
-	port = i.detectPortInDockerfile(filepath.Join("testdata", "Dockerfile", "DockerfilePort3100"))
-	require.Equal(t, 1, len(port))
-	require.Equal(t, 3100, port[0])
-	port = i.detectPortInDockerfile(filepath.Join("testdata", "Dockerfile", "DockerfilePort3100And80"))
-	require.Equal(t, 2, len(port))
-	require.Equal(t, 3100, port[0])
-	require.Equal(t, 80, port[1])
-	port = i.detectPortInDockerfile(filepath.Join("testdata", "Dockerfile", "DockerfileNoPort"))
-	require.Equal(t, 0, len(port))
 }
